backend/requirements.txt
asgiref==3.3.1
Django==3.1.5
django-cors-headers==3.7.0
djangorestframework==3.12.2
pkg-resources==0.0.0
pytz==2020.5
sqlparse==0.4.1
